\name{equiv.regression}
\alias{equiv.regression}

\title{An equivalence between ...
}

\description{
  A long description.
}

\usage{
equiv.regression(yfd, xfd, threshold=0.01)
}

\arguments{
  \item{yfd}{..}
  \item{xfd}{..}
  \item{threshold}{..}
}

\details{
  more details than the description above
}

\value{
  A list with the following objects
  \item{y}{..}
  \item{Xmat}{..}
  \item{cut.value}{..}
}

\references{
Our paper here. (2016). \emph{Submitted to Nucleic Acids Research}.
}

\author{
  Yun Zhang, Juilee Thakar, Xing Qiu
}

%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%% \seealso{
%% }

\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ wMWUTest }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

