\name{FUNNEL.GSEA}
\alias{FUNNEL.GSEA}

\title{The main function (to be edited)
}

\description{
  A long description.
}

\usage{
FUNNEL.GSEA(X, time, geneset, lambda=10^-3.5, rr=rep(1,length(time)), selection_k="FVE", FVE_threshold=0.9, nharm=3, centerfns=FALSE, equiv.threshold=0.01, lam1=0.4, lam2=0.1, alpha.level=0.05)
}

\arguments{
  \item{X}{..}
  \item{time}{..}
  \item{geneset}{..}
  \item{lambda}{..} 
  \item{rr}{..} 
  \item{selection_k}{..} 
  \item{FVE_threshold}{..} 
  \item{nharm}{..} 
  \item{centerfns}{..} 
  \item{equiv.threshold}{..} 
  \item{lam1}{..} 
  \item{lam2}{..} 
  \item{alpha.level}{..} 
}

\details{
  more details than the description above
}

\value{
  A list with the following objects
  \item{pvals}{..}
  \item{weight.list}{..}
  \item{correlation}{..}
  \item{sig.geneset}{..}
}

\references{
Our paper here. (2016). \emph{Submitted to Nucleic Acids Research}.
}

\author{
  Yun Zhang, Juilee Thakar, Xing Qiu
}

%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%% \seealso{
%% }

\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ wMWUTest }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
