\name{plotWeight}
\alias{plotWeight}

\title{Plot ..
}

\description{
  A long description.
}

\usage{
plotWeight(weight.list, geneset.index, main="Weight", ...)
}

\arguments{
  \item{weight.list}{..}
  \item{geneset.index}{..}
  \item{main}{..}
  \item{...}{Other graphical parameters to be passed to function \code{pheatmap()}.}
}

\details{
  more details than the description above
}

\references{
  Our paper here. (2016). \emph{Submitted to Nucleic Acids Research}.
}

\author{
  Yun Zhang, Juilee Thakar, Xing Qiu
}

%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%% \seealso{
%% }

\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ wMWUTest }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

