weightPerGene <- function(weight.list, genesOfInterest)
\name{weightPerGene}
\alias{weightPerGene}

\title{Get weight ...
}

\description{
  A long description.
}

\usage{
weightPerGene(weight.list, genesOfInterest)
}

\arguments{
  \item{weight.list}{..}
  \item{genesOfInterest}{..}
}

\details{
  more details than the description above
}

\value{
  A list with ...
}

\references{
Our paper here. (2016). \emph{Submitted to Nucleic Acids Research}.
}

\author{
  Yun Zhang, Juilee Thakar, Xing Qiu
}

%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%% \seealso{
%% }

\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ wMWUTest }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

